backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
rq==2.1.0
paramiko==3.5.0
requests==2.32.3
reportlab==4.2.5
